improve contextual x_y_plot
